pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tenet"
version = "0.1.0"
description = "Grammar enforcement for Python — tenets of correct library, API, and object usage"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.9",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "anyio[trio]>=4.0",
    "fastapi @ git+https://github.com/ozmalabs/fastapi.git",
    "uvicorn>=0.30",
]

[project.scripts]
tenet = "tenet.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["tenet"]
